ml-service/requirements.txt
# API
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2

# ML & Data
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
scikit-surprise==1.1.4

# Embeddings & Vector DB
sentence-transformers==3.1.1
qdrant-client==1.11.3

# Experiment tracking
mlflow==2.16.2

# Utilities
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
tqdm==4.66.5
